Python Data Analysis
In this data analysis project, I investigated the hypothesis that more goals are scored in women's international soccer matches compared to men's.
   
Using FIFA World Cup data since 2002, I conducted a statistical hypothesis test with a 10% significance level.  

First, I prepared and filtered the datasets for analysis, then created a combined dataset.  

As the goal data was not normally distributed, I utilized the Wilcoxon-Mann-Whitney test.  

With both Pingouin and SciPy libraries, I conducted the test and obtained the p-value. 
 
Finally, based on the significance level, I determined whether to reject or fail to reject the null hypothesis.  

This project enhanced my understanding of statistical hypothesis testing and its application in sports analytics.